+++ examples/auto_option.rb
@@ -12,14 +12,6 @@ class AutoOptionCommand < Clin::Command
   end
 end
 
-if __FILE__ == $PROGRAM_NAME
-  SimpleCommand.parse('-e "Some message 1"').run
-  puts
-  puts '=' * 60
-  puts
-  begin
-    SimpleCommand.parse('--eko="Some message 2"').run
-  rescue Clin::HelpError => e
-    puts e
-  end
-end
+# Run example:
+# AutoOptionCommand.parse('-e "Some message 1"').run
+# AutoOptionCommand.parse('--eko="Some message 2"').run