+++ examples/dispatcher.rb
@@ -23,26 +23,20 @@ class PrintCommand < Clin::Command
   end
 end
 
-if __FILE__ == $PROGRAM_NAME
-  Clin::CommandDispatcher.parse('display "My Message"').run
-  puts
-  puts '=' * 60
-  puts
-  Clin::CommandDispatcher.parse('print "My Message"').run
-  puts
-  puts '=' * 60
-  puts
-  begin
-    Clin::CommandDispatcher.parse('display -h').run
-  rescue Clin::CommandLineError => e
-    puts e
-  end
-  puts
-  puts '=' * 60
-  puts
-  begin
-    Clin::CommandDispatcher.parse('-h')
-  rescue Clin::CommandLineError => e
-    puts e
+# Simple command Example
+class PrintAltCommand < Clin::Command
+  arguments 'print <message>'
+
+  description 'Print the given message alternative'
+  prioritize
+
+  def run
+    puts "Print alt: '#{params[:message]}'"
   end
 end
+
+# Example of run:
+# Clin::CommandDispatcher.parse('display "My Message"').run
+# Clin::CommandDispatcher.parse('print "My Message"').run
+# Clin::CommandDispatcher.parse('display -h').run
+# Clin::CommandDispatcher.parse('-h')